bugfixes, BUT
