+++ unstable/httpy/httpy.py
@@ -40,6 +40,8 @@ import queue  # to communicate between threads
 from . import http2
 from .utils import *
 from .errors import *
+from .status import *
+
 try:
     import chardet  # to detect charsets
 except ImportError:
@@ -81,294 +83,10 @@ WEBSOCKET_CLOSE_CODES = {
     1014: "Bad Gateway",
     1015: "TLS handshake",
 }
-STATUS_CODES = {
-    "100": {
-        "code": "100",
-        "message": "Continue",
-        "description": "indicates that the initial part of a request has been received and has not yet been rejected by the server.",
-    },
-    "101": {
-        "code": "101",
-        "message": "Switching Protocols",
-        "description": "indicates that the server understands and is willing to comply with the client's request, via the Upgrade header field, for a change in the application protocol being used on this connection.",
-    },
-    "200": {
-        "code": "200",
-        "message": "OK",
-        "description": "indicates that the request has succeeded.",
-    },
-    "201": {
-        "code": "201",
-        "message": "Created",
-        "description": "indicates that the request has been fulfilled and has resulted in one or more new resources being created.",
-    },
-    "202": {
-        "code": "202",
-        "message": "Accepted",
-        "description": "indicates that the request has been accepted for processing, but the processing has not been completed.",
-    },
-    "203": {
-        "code": "203",
-        "message": "Non-Authoritative Information",
-        "description": "indicates that the request was successful but the enclosed payload has been modified from that of the origin server's 200 (OK) response by a transforming proxy.",
-    },
-    "204": {
-        "code": "204",
-        "message": "No Content",
-        "description": "indicates that the server has successfully fulfilled the request and that there is no additional content to send in the response payload body.",
-    },
-    "205": {
-        "code": "205",
-        "message": "Reset Content",
-        "description": "indicates that the server has fulfilled the request and desires that the user agent reset the document view, which caused the request to be sent, to its original state as received from the origin server.",
-    },
-    "206": {
-        "code": "206",
-        "message": "Partial Content",
-        "description": "indicates that the server is successfully fulfilling a range request for the target resource by transferring one or more parts of the selected representation that correspond to the satisfiable ranges found in the requests's Range header field.",
-    },
-    "300": {
-        "code": "300",
-        "message": "Multiple Choices",
-        "description": "indicates that the target resource has more than one representation, each with its own more specific identifier, and information about the alternatives is being provided so that the user (or user agent) can select a preferred representation by redirecting its request to one or more of those identifiers.",
-    },
-    "301": {
-        "code": "301",
-        "message": "Moved Permanently",
-        "description": "indicates that the target resource has been assigned a new permanent URI and any future references to this resource ought to use one of the enclosed URIs.",
-    },
-    "302": {
-        "code": "302",
-        "message": "Found",
-        "description": "indicates that the target resource resides temporarily under a different URI.",
-    },
-    "303": {
-        "code": "303",
-        "message": "See Other",
-        "description": "indicates that the server is redirecting the user agent to a different resource, as indicated by a URI in the Location header field, that is intended to provide an indirect response to the original request.",
-    },
-    "304": {
-        "code": "304",
-        "message": "Not Modified",
-        "description": "indicates that a conditional GET request has been received and would have resulted in a 200 (OK) response if it were not for the fact that the condition has evaluated to false.",
-    },
-    "305": {
-        "code": "305",
-        "message": "Use Proxy",
-        "description": "*deprecated*",
-    },
-    "307": {
-        "code": "307",
-        "message": "Temporary Redirect",
-        "description": "indicates that the target resource resides temporarily under a different URI and the user agent MUST NOT change the request method if it performs an automatic redirection to that URI.",
-    },
-    "400": {
-        "code": "400",
-        "message": "Bad Request",
-        "description": "indicates that the server cannot or will not process the request because the received syntax is invalid, nonsensical, or exceeds some limitation on what the server is willing to process.",
-    },
-    "401": {
-        "code": "401",
-        "message": "Unauthorized",
-        "description": "indicates that the request has not been applied because it lacks valid authentication credentials for the target resource.",
-    },
-    "402": {
-        "code": "402",
-        "message": "Payment Required",
-        "description": "*reserved*",
-    },
-    "403": {
-        "code": "403",
-        "message": "Forbidden",
-        "description": "indicates that the server understood the request but refuses to authorize it.",
-    },
-    "404": {
-        "code": "404",
-        "message": "Not Found",
-        "description": "indicates that the origin server did not find a current representation for the target resource or is not willing to disclose that one exists.",
-    },
-    "405": {
-        "code": "405",
-        "message": "Method Not Allowed",
-        "description": "indicates that the method specified in the request-line is known by the origin server but not supported by the target resource.",
-    },
-    "406": {
-        "code": "406",
-        "message": "Not Acceptable",
-        "description": "indicates that the target resource does not have a current representation that would be acceptable to the user agent, according to the proactive negotiation header fields received in the request, and the server is unwilling to supply a default representation.",
-    },
-    "407": {
-        "code": "407",
-        "message": "Proxy Authentication Required",
-        "description": "is similar to 401 (Unauthorized), but indicates that the client needs to authenticate itself in order to use a proxy.",
-    },
-    "408": {
-        "code": "408",
-        "message": "Request Timeout",
-        "description": "indicates that the server did not receive a complete request message within the time that it was prepared to wait.",
-    },
-    "409": {
-        "code": "409",
-        "message": "Conflict",
-        "description": "indicates that the request could not be completed due to a conflict with the current state of the resource.",
-    },
-    "410": {
-        "code": "410",
-        "message": "Gone",
-        "description": "indicates that access to the target resource is no longer available at the origin server and that this condition is likely to be permanent.",
-    },
-    "411": {
-        "code": "411",
-        "message": "Length Required",
-        "description": "indicates that the server refuses to accept the request without a defined Content-Length.",
-    },
-    "412": {
-        "code": "412",
-        "message": "Precondition Failed",
-        "description": "indicates that one or more preconditions given in the request header fields evaluated to false when tested on the server.",
-    },
-    "413": {
-        "code": "413",
-        "message": "Payload Too Large",
-        "description": "indicates that the server is refusing to process a request because the request payload is larger than the server is willing or able to process.",
-    },
-    "414": {
-        "code": "414",
-        "message": "URI Too Long",
-        "description": "indicates that the server is refusing to service the request because the request-target is longer than the server is willing to interpret.",
-    },
-    "415": {
-        "code": "415",
-        "message": "Unsupported Media Type",
-        "description": "indicates that the origin server is refusing to service the request because the payload is in a format not supported by the target resource for this method.",
-    },
-    "416": {
-        "code": "416",
-        "message": "Range Not Satisfiable",
-        "description": "indicates that none of the ranges in the request's Range header field overlap the current extent of the selected resource or that the set of ranges requested has been rejected due to invalid ranges or an excessive request of small or overlapping ranges.",
-    },
-    "417": {
-        "code": "417",
-        "message": "Expectation Failed",
-        "description": "indicates that the expectation given in the request's Expect header field could not be met by at least one of the inbound servers.",
-    },
-    "418": {
-        "code": "418",
-        "message": "I'm a teapot",
-        "description": "Any attempt to brew coffee with a teapot should result in the error code 418 I'm a teapot.",
-    },
-    "426": {
-        "code": "426",
-        "message": "Upgrade Required",
-        "description": "indicates that the server refuses to perform the request using the current protocol but might be willing to do so after the client upgrades to a different protocol.",
-    },
-    "500": {
-        "code": "500",
-        "message": "Internal Server Error",
-        "description": "indicates that the server encountered an unexpected condition that prevented it from fulfilling the request.",
-    },
-    "501": {
-        "code": "501",
-        "message": "Not Implemented",
-        "description": "indicates that the server does not support the functionality required to fulfill the request.",
-    },
-    "502": {
-        "code": "502",
-        "message": "Bad Gateway",
-        "description": "indicates that the server, while acting as a gateway or proxy, received an invalid response from an inbound server it accessed while attempting to fulfill the request.",
-    },
-    "503": {
-        "code": "503",
-        "message": "Service Unavailable",
-        "description": "indicates that the server is currently unable to handle the request due to a temporary overload or scheduled maintenance, which will likely be alleviated after some delay.",
-    },
-    "504": {
-        "code": "504",
-        "message": "Gateway Time-out",
-        "description": "indicates that the server, while acting as a gateway or proxy, did not receive a timely response from an upstream server it needed to access in order to complete the request.",
-    },
-    "505": {
-        "code": "505",
-        "message": "HTTP Version Not Supported",
-        "description": "indicates that the server does not support, or refuses to support, the protocol version that was used in the request message.",
-    },
-    "102": {
-        "code": "102",
-        "message": "Processing",
-        "description": "is an interim response used to inform the client that the server has accepted the complete request, but has not yet completed it.",
-    },
-    "207": {
-        "code": "207",
-        "message": "Multi-Status",
-        "description": "provides status for multiple independent operations.",
-    },
-    "226": {
-        "code": "226",
-        "message": "IM Used",
-        "description": "The server has fulfilled a GET request for the resource, and the response is a representation of the result of one or more instance-manipulations applied to the current instance.",
-    },
-    "308": {
-        "code": "308",
-        "message": "Permanent Redirect",
-        "description": "The target resource has been assigned a new permanent URI and any future references to this resource outght to use one of the enclosed URIs. [...] This status code is similar to 301 Moved Permanently (Section 7.3.2 of rfc7231), except that it does not allow rewriting the request method from POST to GET.",
-    },
-    "422": {
-        "code": "422",
-        "message": "Unprocessable Entity",
-        "description": "means the server understands the content type of the request entity (hence a 415(Unsupported Media Type) status code is inappropriate), and the syntax of the request entity is correct (thus a 400 (Bad Request) status code is inappropriate) but was unable to process the contained instructions.",
-    },
-    "423": {
-        "code": "423",
-        "message": "Locked",
-        "description": "means the source or destination resource of a method is locked.",
-    },
-    "424": {
-        "code": "424",
-        "message": "Failed Dependency",
-        "description": "means that the method could not be performed on the resource because the requested action depended on another action and that action failed.",
-    },
-    "428": {
-        "code": "428",
-        "message": "Precondition Required",
-        "description": "indicates that the origin server requires the request to be conditional.",
-    },
-    "429": {
-        "code": "429",
-        "message": "Too Many Requests",
-        "description": "indicates that the user has sent too many requests in a given amount of time (rate limiting).",
-    },
-    "431": {
-        "code": "431",
-        "message": "Request Header Fields Too Large",
-        "description": "indicates that the server is unwilling to process the request because its header fields are too large.",
-    },
-    "451": {
-        "code": "451",
-        "message": "Unavailable For Legal Reasons",
-        "description": "This status code indicates that the server is denying access to the resource in response to a legal demand.",
-    },
-    "506": {
-        "code": "506",
-        "message": "Variant Also Negotiates",
-        "description": "indicates that the server has an internal configuration error: the chosen variant resource is configured to engage in transparent content negotiation itself, and is therefore not a proper end point in the negotiation process.",
-    },
-    "507": {
-        "code": "507",
-        "message": "Insufficient Storage",
-        "description": "means the method could not be performed on the resource because the server is unable to store the representation needed to successfully complete the request.",
-    },
-    "511": {
-        "code": "511",
-        "message": "Network Authentication Required",
-        "description": "indicates that the client needs to authenticate to gain network access.",
-    },
-}
 
 context = ssl._create_default_https_context()
 context.set_alpn_protocols(["http/1.1"])
 schemes = {"http": 80, "https": 443}
-
-
 
 
 def _mk2l(original):
@@ -420,33 +138,6 @@ def mask(data, mask):
     return r
 
 
-class Status(int):
-    """
-    Creates HTTP status from string.
-
-    :param statstring: string to parse
-    """
-
-    def __init__(self, statstring):
-        _, self.status, self.reason = STATUSPATTERN.search(statstring).groups()
-        self.status = int(self.status)
-        try:
-            self.codes_entry = STATUS_CODES[str(self.status)]
-        except KeyError:
-            self.codes_entry = {
-                "code": str(self.status),
-                "message": "Unknown",
-                "description": "Unknown",
-            }
-        self.__dict__.update(self.codes_entry)
-
-        self.reason = self.reason.decode()
-
-    def __new__(cls, statstring):
-        _, status, reason = STATUSPATTERN.search(statstring).groups()
-        return super(Status, cls).__new__(cls, status)
-
-
 class _Debugger:
     """
     Debugger
@@ -495,59 +186,6 @@ class _Debugger:
     ok = debugging_method("\033[92;1m[OK]", "\033[0m")
     warn = debugging_method("\033[93;1m[WARN]", "\033[0m")
     error = debugging_method("\033[31;1m[ERROR]", "\033[0m")
-
-
-def _find(key, d):
-    for i in d:
-        if i.lower() == key.lower():
-            return i
-    return key
-
-
-class CaseInsensitiveDict(dict):
-    """Case insensitive subclass of dictionary"""
-
-    def __init__(self, data):
-        self.lowercase = {force_string(k).lower(): v for k, v in dict(data).items()}
-        self.original = data
-        super().__init__(self.original)
-
-    def __contains__(self, item):
-        return (force_string(item).lower() in self.lowercase) | (
-            force_string(item) in self.original
-        )
-
-    def __getitem__(self, item):
-        try:
-            return self.lowercase[force_string(item).lower()]
-        except KeyError:
-            return self.original[force_string(item)]
-
-    def __setitem__(self, item, val):
-        self.lowercase[force_string(item).lower()] = val
-        self.original[_find(item, self.original)] = val
-        super().__init__(self.original)  # remake??
-
-    def get(self, key, default=None):
-        if key in self:
-            return self[key]
-        return default
-
-    def update(self, d):
-        for key in d:
-            self[key] = d[key]
-
-    def __iter__(self):
-        return iter(self.original)
-
-    def keys(self):
-        return self.original.keys()
-
-    def values(self):
-        return self.original.values()
-
-    def items(self):
-        return self.original.items()
 
 
 def _binappendstr(s):
@@ -1348,7 +986,9 @@ class ConnectionPool:
 
     def __setitem__(self, host, connection):
         host, port = host
-        if (connection.http2 and connection._sock.sock.fileno()==-1) or( not connection.http2 and connection._sock.fileno() == -1):
+        if (connection.http2 and connection._sock.sock.fileno() == -1) or (
+            not connection.http2 and connection._sock.fileno() == -1
+        ):
             raise ConnectionClosedError("Connection closed by host")
         self.connections[host, port] = connection
 
@@ -1359,8 +999,7 @@ class ConnectionPool:
             raise
             del self.connections[host]
 
-            raise ConnectionClosedError("Connection closed by host")
-        if sock.fileno() == -1:
+        if getattr(sock, "fileno", sock.sock.fileno)() == -1:
             del self.connections[host]
             raise ConnectionClosedError("Connection closed by host")
         return sock
@@ -1493,8 +1132,6 @@ def _generate_boundary():
         + "".join(random.choices(string.ascii_letters + string.digits, k=10)).encode()
         + b"\r\n"
     )
-
-
 
 
 def get_content_type(data):
@@ -1655,7 +1292,7 @@ def create_connection(host, port, last_response, is_http2):
     try:
         if is_http2:
             debugger.info("instancing http2 connection")
-            conn = http2.Connection(host,port,debugger)
+            conn = http2.Connection(host, port, debugger)
             conn.start()
         else:
             debugger.info("calling socket.create_connection")
@@ -1680,6 +1317,8 @@ def create_connection(host, port, last_response, is_http2):
         raise  # Added in 1.1.1
     pool[host, port] = Connection(conn, keep_alive.timeout, keep_alive.max, is_http2)
     return conn, False
+
+
 def generate_session_id():
     return "".join(random.choices(string.hexdigits, k=16))
 
@@ -1795,10 +1434,12 @@ class HTTP11(ProtoVersion):
     sender = HTTP11Sender
     recver = HTTP11Recver()
 
+
 class HTTP2(ProtoVersion):
     version = "2"
     sender = http2.proto.HTTP2Sender
     recver = http2.proto.HTTP2Recver()
+
 
 class Session:
     def __init__(self, session_id=None, path=None, name=None):
@@ -1860,7 +1501,7 @@ def _raw_request(
     disabled_headers=[],
     force_keep_alive=True,
 ):
-    is_http2 = http_version=="2"
+    is_http2 = http_version == "2"
     method = method.upper()
     cache = Cache(base_dir / "sites")
     jar = CookieJar(base_dir / "cj")
@@ -1948,10 +1589,14 @@ def _raw_request(
         if cf:
             cf.add_header(defhdr)
         proto = proto_versions[http_version]
-        proto.send_request(sock, method, defhdr, data, path, debug)
-        status, resp_headers, decoded_body, body = proto.recv_response(
-            sock, debug, timeout
-        )
+        ret_val = proto.send_request(sock, method, defhdr, data, path, debug)
+        if is_http2:
+            args = (sock, ret_val)
+        else:
+            args = (sock, debug, timeout)
+
+        status, resp_headers, decoded_body, body = proto.recv_response(*args)
+
         if status == 304:
             return Response.cacheload(cf)
         if "set-cookie" in resp_headers:
@@ -2001,9 +1646,7 @@ def _raw_request(
     if headers.get("connection") == "keep-alive":
         pool.connections[
             host, port
-        ]._sock = (
-            sock
-        )  # Fix bug #23 -- New  connections in keep-alive mode slowing down requests
+        ]._sock = sock  # Fix bug #23 -- New  connections in keep-alive mode slowing down requests
     end_time = time.time()
     elapsed_time = end_time - start_time
